Prêt pour le pull. Ajout des classes, interfaces et màj du BddContext.

+++ EasyTrain_P2Gr1/EasyTrain_P2Gr1/Models/Cours.cs
@@ -5,7 +5,7 @@
         public int Id { get; set; }
         public string Titre { get; set; }
         public int NbParticipants { get; set; }
-        public float Prix { get; set; }
+        public double Prix { get; set; }
         public virtual Coach Coach { get; set; }
         public virtual Salle Salle { get; set; }
